minor mixes
