+++ .changeset/small-mirrors-trade.md
@@ -1,3 +1,4 @@
 ---
-"@finos/legend-graph": patch
+'@finos/legend-graph': patch
+'@finos/legend-query': patch
 ---
