+++ packages/legend-application/src/stores/ValueSpecificationUtils.ts
@@ -17,33 +17,11 @@
 import {
   type PureModel,
   type PRIMITIVE_TYPE,
-  type Multiplicity,
-  MULTIPLICITY_INFINITE,
   PrimitiveInstanceValue,
   TYPICAL_MULTIPLICITY_TYPE,
   GenericTypeExplicitReference,
   GenericType,
 } from '@finos/legend-graph';
-
-export const getMultiplicityDescription = (
-  multiplicity: Multiplicity,
-): string => {
-  if (multiplicity.lowerBound === multiplicity.upperBound) {
-    return `[${multiplicity.lowerBound.toString()}] - Must have exactly ${multiplicity.lowerBound.toString()} value(s)`;
-  } else if (
-    multiplicity.lowerBound === 0 &&
-    multiplicity.upperBound === undefined
-  ) {
-    return `[${MULTIPLICITY_INFINITE}] - May have many values`;
-  }
-  return `[${multiplicity.lowerBound}..${
-    multiplicity.upperBound ?? MULTIPLICITY_INFINITE
-  }] - ${
-    multiplicity.upperBound
-      ? `Must have from ${multiplicity.lowerBound} to ${multiplicity.upperBound} value(s)`
-      : `Must have at least ${multiplicity.lowerBound} values(s)`
-  }`;
-};
 
 export const buildPrimitiveInstanceValue = (
   graph: PureModel,
