+++ packages/legend-query/package.json
@@ -53,7 +53,6 @@
     "@finos/legend-server-depot": "workspace:*",
     "@finos/legend-shared": "workspace:*",
     "@testing-library/react": "13.3.0",
-    "@types/papaparse": "5.3.2",
     "@types/react": "18.0.10",
     "@types/react-dom": "18.0.5",
     "@types/react-router-dom": "5.3.2",
@@ -62,7 +61,6 @@
     "mobx": "6.6.0",
     "mobx-react-lite": "3.4.0",
     "monaco-editor": "0.33.0",
-    "papaparse": "5.3.2",
     "react": "18.1.0",
     "react-dnd": "16.0.1",
     "react-dnd-html5-backend": "16.0.1",
